+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Miscellaneous
+- Fix tests style
+
 ## [1.0.2] - 2018-02-17
 
 ### Fixed